CNNs/requirements.txt
matplotlib==3.5.1
numpy==1.22.2
Pillow==8.4.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
tensorflow-gpu==2.3.0
split-folders==0.5.1
opencv-python==4.5.5.62

